Move S3 and boto fixtures into helper functions
