+++ tests/sagemaker/mock/test_sagemaker_service_mock.py
@@ -3,14 +3,8 @@ import os
 import boto3
 import pytest
 
+from tests.helper_functions import set_boto_credentials
 from tests.sagemaker.mock import mock_sagemaker
-
-
-@pytest.fixture(scope='session', autouse=True)
-def set_boto_credentials():
-    os.environ["AWS_ACCESS_KEY_ID"] = "NotARealAccessKey"
-    os.environ["AWS_SECRET_ACCESS_KEY"] = "NotARealSecretAccessKey"
-    os.environ["AWS_SESSION_TOKEN"] = "NotARealSessionToken"
 
 
 @pytest.fixture
